Added swag, arduino unly updates if volume is new

+++ Android/app/src/main/java/swag/theokanning/airhorn/bluetooth/AirhornConnectionListener.java
@@ -1,7 +1,7 @@
 package swag.theokanning.airhorn.bluetooth;
 
 public interface AirhornConnectionListener {
-    void onVolumeChanged(byte volume);
+    void onVolumeChanged(float volume);
 
     void onConnect();
